tercera entrega A

+++ src/controllers/products.controller.js
@@ -29,15 +29,29 @@ class ProductController {
         } 
     }
 
-    putProduct = async (req,res) => {
+    putProduct = async (req, res) => {
         try {
-            let id = req.params.id
-            let updProd = req.body
-            res.send(await serviceProduct.updateProduct(id, updProd))
+            let id = req.params.id;
+            let updProd = req.body;
+
+            if (req.user.rol === 'admin') {
+                res.send(await serviceProduct.updateProduct(id, updProd));
+            } else {
+                const product = await serviceProduct.getProductById(id);
+                if (product && product.owner === req.user.email) {
+                    res.send(await serviceProduct.updateProduct(id, updProd));
+                } else {
+                    throw new CustomError({
+                        name: 'Permission error',
+                        message: 'No tienes permisos para modificar este producto',
+                        code: EErrors.PERMISSION_ERROR,
+                    });
+                }
+            }
         } catch (error) {
-            logger.log(error)
+            logger.error('Error al modificar el producto:', error);
+            res.status(500).json({ error: 'Error al modificar el producto' });
         }
-        
     }
 
     getProductById = async (req, res) => {
@@ -53,30 +67,49 @@ class ProductController {
 
     deleteProduct = async (req, res) => {
         try {
-            let id = req.params.id
-            res.send(await serviceProduct.deleteProduct(id))
+            let id = req.params.id;
+
+            if (req.user.rol === 'admin') {
+                res.send(await serviceProduct.deleteProduct(id));
+            } else {
+                const product = await serviceProduct.getProductById(id);
+                if (product && product.owner === req.user.email) {
+                    res.send(await serviceProduct.deleteProduct(id));
+                } else {
+                    throw new CustomError({
+                        name: 'Permission error',
+                        message: 'No tienes permisos para eliminar este producto',
+                        code: EErrors.PERMISSION_ERROR,
+                    });
+                }
+            }
         } catch (error) {
-            logger.log(error)
+            logger.error('Error al eliminar el producto:', error);
+            res.status(500).json({ error: 'Error al eliminar el producto' });
         }
-        
     }
 
     postProduct = async (req, res, next) => {
         try {
-            let newProduct = req.body
-            if(newProduct == null){
+            let newProduct = req.body;
+
+            const productOwner = newProduct.owner || 'admin';
+
+            newProduct.owner = productOwner;
+
+            if (newProduct == null) {
                 CustomError.createError({
                     name: 'Product creation error',
                     cause: generateProductErrorInfo(newProduct),
                     message: 'Error trying to create a product, invalid request',
-                    code: EErrors.INVALID_REQUEST
-                })
+                    code: EErrors.INVALID_REQUEST,
+                });
             }
-            res.send(await serviceProduct.addProduct(newProduct))
+
+            res.send(await serviceProduct.addProduct(newProduct));
         } catch (error) {
-            next(error)
+            next(error);
         }
-        
     }
 
 }
